pyproject.toml
[tool.poetry]
name = "calchecker"
version = "0.1.0"
description = ""
authors = ["Martim Lobao <m@rtimlobao.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
cryptography = "^42.0.7"
icalendar = "^5.0.12"
python-dotenv = "^1.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = ">=24.1.1"
coverage = ">=7.4.1"
docformatter = ">=1.7.5"
hypothesis = ">=6.97.3"
ipython = ">=8.20.0"
mypy = ">=1.8.0"
pylint = ">=3.0.3"
pytest = ">=8.0.0"
pytest-cov = ">=4.1.0"
ruff = ">=0"
types-beautifulsoup4 = ">=4.12.0.5"
types-pytz = ">=2023.4.0.0"
types-requests = ">=2.31.0.1"
yamllint = ">=1.33.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 99
preview = true

[tool.docformatter]
recursive = true
style = "google"
wrap-descriptions = 80
wrap-summaries = 80
